+++ optd-cost-model-perf/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod dbms;
 pub mod shell;
 pub mod tpch;
-pub mod tpch_q6;
